refactor(e2e): move API tests to api_spec.js
